==> 01_data_ingestion.ipynb <==
import pandas as pd
import boto3
import sagemaker
import matplotlib.pyplot as plt
import os
import numpy as np




!aws s3 ls s3://oneclickdata/aisles/

!aws s3 ls s3://oneclickdata/departments/

!aws s3 ls s3://oneclickdata/prior/

!aws s3 ls s3://oneclickdata/train/

!aws s3 ls s3://oneclickdata/orders/

!aws s3 ls s3://oneclickdata/products/

s3_public_path_csv = "s3://oneclickdata"

s3_public_path_csv

sess= sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name
account_id = boto3.client("sts").get_caller_identity().get("Account")

sm = boto3.Session().client(service_name="sagemaker", region_name=region)

s3_private_path_csv = "s3://{}/oneclickdata".format(bucket)
print(s3_private_path_csv)



!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "aisles/aisles.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "departments/departments.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "prior/order_products__prior.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "train/order_products__train.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "orders/orders.csv"
!aws s3 cp --recursive $s3_public_path_csv/ $s3_private_path_csv/ --exclude "*" --include "products/products.csv"

!aws s3 ls $s3_private_path_csv/aisles/

from IPython.core.display import display, HTML

display(
    HTML(
        '<b>Review <a target="blank" href="https://s3.console.aws.amazon.com/s3/buckets/sagemaker-{}-{}/oneclickdata/?region={}&tab=overview">S3 Bucket</a></b>'.format(
            region, account_id, region
        )
    )
)
