==> 03_createathena_table_query.ipynb <==
import pandas as pd
import boto3
import sagemaker

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

database_name = "oneclickdb"

from pyathena import connect

# Set S3 staging directory -- this is a temporary directory used for Athena queries
s3_staging_dir = "s3://{0}/athena/staging".format(bucket)




%store -r ingest_create_athena_db_passed




private_s3_path_csv1 = "s3://{}/oneclickdata/aisles/".format(bucket)
print(private_s3_path_csv1)



private_s3_path_csv2 = "s3://{}/oneclickdata/departments/".format(bucket)
print(private_s3_path_csv2)

private_s3_path_csv3= "s3://{}/oneclickdata/prior/".format(bucket)
print(private_s3_path_csv3)

private_s3_path_csv4= "s3://{}/oneclickdata/train/".format(bucket)
print(private_s3_path_csv4)

private_s3_path_csv5 = "s3://{}/oneclickdata/orders/".format(bucket)
print(private_s3_path_csv5)

private_s3_path_csv6 = "s3://{}/oneclickdata/products/".format(bucket)
print(private_s3_path_csv6)

# Set Athena parameters
table_1 = "aisles_tab"
table_2= "departments_tab"
table_3= "order_prior_tab"
table_4= "order_train_tab"
table_5= "orders_tab"
table_6= "products_tab"

conn = connect(region_name=region, s3_staging_dir=s3_staging_dir)

# SQL statement to execute aisles table
statement1 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         aisle_id int,
         aisle string
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table_1, private_s3_path_csv1
)


pd.read_sql(statement1, conn)


# drop1= "DROP TABLE oneclickaws.aisles"
# pd.read_sql(drop1, conn)

# drop2= "DROP TABLE oneclickaws.departments"
# pd.read_sql(drop2, conn)

# drop3= "DROP TABLE oneclickaws.order_products__prior"
# pd.read_sql(drop3, conn)

# drop4= "DROP TABLE oneclickaws.order_products__train"
# pd.read_sql(drop4, conn)

# drop5= "DROP TABLE oneclickaws.orders"
# pd.read_sql(drop5, conn)

# drop5= "DROP TABLE oneclickaws.products"
# pd.read_sql(drop5, conn)

# SQL statement to execute department table
statement2 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         department_id int,
         department string
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table_2, private_s3_path_csv2
)


pd.read_sql(statement2, conn)

# SQL statement to execute order_products__prior table
statement3 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         order_id int,
         product_id int,
         add_to_cart_order int,
         reordered string
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table_3, private_s3_path_csv3
)


pd.read_sql(statement3, conn)

# SQL statement to execute order_products__train table
statement4 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         order_id int,
         product_id int,
         add_to_cart_order int,
         reordered string
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table_4, private_s3_path_csv4
)


pd.read_sql(statement4, conn)

# SQL statement to execute orders table
statement5 = """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
         order_id int,
         user_id int,
         eval_set string,
         order_number int,
         order_dow string,
         order_hour_of_day string,
         days_since_prior_order float
       
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table_5, private_s3_path_csv5
)


pd.read_sql(statement5, conn)

# SQL statement to execute products table
statement6= """CREATE EXTERNAL TABLE IF NOT EXISTS {}.{}(
          product_id int,
         product_name string,
         aisle_id int,
         department_id int
) ROW FORMAT DELIMITED
FIELDS TERMINATED BY ',' 
LINES TERMINATED BY '\n'
LOCATION '{}'
TBLPROPERTIES ('skip.header.line.count'='1')""".format(
    database_name, table_6, private_s3_path_csv6
)


pd.read_sql(statement6, conn)

statement = "SHOW TABLES in {}".format(database_name)

df_show = pd.read_sql(statement, conn)
df_show.head(6)

test_statement = """SELECT * FROM {}.{} LIMIT 5 """.format(
    database_name, table_1
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(5)

test_statement = """SELECT * FROM {}.{} LIMIT 5 """.format(
    database_name, table_2
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(5)

test_statement = """SELECT * FROM {}.{} LIMIT 10 """.format(
    database_name, table_3
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(10)

test_statement = """SELECT * FROM {}.{} LIMIT 10 """.format(
    database_name, table_4
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(10)

test_statement = """SELECT * FROM {}.{} LIMIT 10 """.format(
    database_name, table_5
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(10)

test_statement = """SELECT * FROM {}.{} LIMIT 10 """.format(
    database_name, table_6
)

print(test_statement)
df_show = pd.read_sql(test_statement, conn)
df_show.head(10)
