==> 07_feature_extraction.ipynb <==
import pandas as pd
import numpy as np

import sagemaker
import boto3

sess = sagemaker.Session()
bucket = sess.default_bucket()
role = sagemaker.get_execution_role()
region = boto3.Session().region_name

sm = boto3.Session().client(service_name="sagemaker", region_name=region)
s3 = boto3.Session().client(service_name="s3", region_name=region)

# Data overview

aisles = pd.read_csv("./data-clarify/aisles.csv")
departments = pd.read_csv("./data-clarify/departments.csv")
orders = pd.read_csv("./data-clarify/orders.csv")
order_products_prior = pd.read_csv("./data-clarify/order_products__prior.csv")
order_products_train = pd.read_csv("./data-clarify/order_products__train.csv")
products = pd.read_csv("./data-clarify/products.csv")

## Merging train data with orders data and products data

train_df = order_products_train.merge(orders, on ='order_id', how='inner')
train_df = train_df.merge(products, on = 'product_id', how = 'left')
train_df.head()

## Features creation

Calculating how many times a user buy the product

train_df['user_buy_product_times'] = train_df.groupby(['user_id', 'product_id']).cumcount() + 1
train_df.head()


## Product level features

(1) Product's average add-to-cart-order

(2) Total times the product was ordered

(3) Total times the product was reordered

(4) Reorder percentage of a product

(5) Total unique users of a product




prod_features = train_df.groupby('product_id').agg(mean_add_to_cart_order=pd.NamedAgg(column='add_to_cart_order',aggfunc='mean'),\
                                                 total_orders=pd.NamedAgg(column='reordered',aggfunc='count'),\
                                                 total_reorders=pd.NamedAgg(column='reordered',aggfunc='sum'),\
                                                 reorder_percentage=pd.NamedAgg(column='reordered',aggfunc='mean'),\
                                                 unique_users=pd.NamedAgg(column='user_id',aggfunc=lambda x: x.nunique()),\
                                                 order_first_time_total_cnt=pd.NamedAgg(column='user_buy_product_times',aggfunc=lambda x: sum(x==1)))
                            
                                                  

prod_features.reset_index(inplace = True)




prod_features.head()

## Aisle and department features


(7) Reorder percentage, Total orders and reorders of a product aisle

(8) Mean and std of aisle add-to-cart-order

(9) Aisle unique users



aisle_features= train_df.groupby('aisle_id').agg(aisle_mean_add_to_cart_order=pd.NamedAgg(column='add_to_cart_order',aggfunc='mean'),\
                                              aisle_std_add_to_cart_order=pd.NamedAgg(column='add_to_cart_order',aggfunc='std'),\
                                              aisle_total_orders=pd.NamedAgg(column='reordered',aggfunc='count'),\
                                              aisle_total_reorders=pd.NamedAgg(column='reordered',aggfunc='sum'),\
                                              aisle_reorder_percentage=pd.NamedAgg(column='reordered',aggfunc='mean'),\
                                              aisle_unique_users=pd.NamedAgg(column='user_id',aggfunc=lambda x: x.nunique()))
                                              
aisle_features.reset_index(inplace = True)                                            

aisle_features.head()

#### features
(10) Reorder percentage, Total orders and reorders of a product department

(11) Mean and std of department add-to-cart-order

(12) Department unique users

dpt_features= train_df.groupby('department_id').agg(department_mean_add_to_cart_order=pd.NamedAgg(column='add_to_cart_order',aggfunc='mean'),\
                                                 department_std_add_to_cart_order=pd.NamedAgg(column='add_to_cart_order',aggfunc='std'),\
                                                 department_total_orders=pd.NamedAgg(column='reordered',aggfunc='count'),\
                                                department_total_reorders=pd.NamedAgg(column='reordered',aggfunc='sum'),\
                                                 department_reorder_percentage=pd.NamedAgg(column='reordered',aggfunc='mean'),\
                                                  department_unique_users=pd.NamedAgg(column='user_id',aggfunc=lambda x: x.nunique()))
                                                 
dpt_features.reset_index(inplace = True)                                                 

dpt_features.head()

#### features

(13) Binary encoding of aisle feature

(14) Binary encoding of department feature



# combine all previous dataframes
prod_features = prod_features.merge(products, on = 'product_id', how = 'left')
prod_features = prod_features.merge(aisle_features, on = 'aisle_id', how = 'left')
prod_features = prod_features.merge(aisles, on = 'aisle_id', how = 'left')
prod_features = prod_features.merge(dpt_features, on = 'department_id', how = 'left')
prod_features = prod_features.merge(departments, on = 'department_id', how = 'left')
prod_features.head()



prod_features.drop(['product_name', 'aisle_id', 'department_id'], axis = 1, inplace = True)
prod_features.head()

prod_features.shape

prod_features.dtypes

## User level features 

(15) Total orders by a user

(16) Total products user has bought

(17) Total unique products user has bought

(18) user's total reordered products

(19) User's overall reorder percentage



user_features = train_df.groupby('user_id').agg(total_orders_by_user=pd.NamedAgg(column='order_number',aggfunc=lambda x: x.nunique()),\
                                             total_products_by_user=pd.NamedAgg(column='product_id',aggfunc='count'),\
                                            total_unique_product_by_user=pd.NamedAgg(column='product_id',aggfunc=lambda x: x.nunique()),\
                                              total_reorders_by_user=pd.NamedAgg(column='reordered',aggfunc='sum'),\
                                          reorder_propotion_by_user=pd.NamedAgg(column='reordered',aggfunc='mean'))
                                             

user_features.reset_index(inplace = True)  

user_features.head()

#### features

(20) Average order size of a user

(21) User's mean of reordered items of all orders



user_features2 = train_df.groupby(['user_id','order_number']).agg(average_order_size=pd.NamedAgg(column='reordered',aggfunc='count'),\
                                                                reorder_in_order=pd.NamedAgg(column='reordered',aggfunc='mean'))
user_features2.reset_index(inplace = True)  
user_features2.head(20)

user_features3 = user_features2.groupby('user_id').agg({'average_order_size' : 'mean', 
                                   'reorder_in_order':'mean'})
user_features3.reset_index(inplace = True)  
user_features3.head()


user_features = user_features.merge(user_features3, on = 'user_id', how = 'left')
user_features.head()

path = "./data-clarify/product_features.pkl"

prod_features.to_pickle(path)

path = "./data-clarify/user_features.pkl"

user_features.to_pickle(path)

df = pd.read_pickle("./data-clarify/product_features.pkl")
df.head()

df = pd.read_pickle("./data-clarify/user_features.pkl")
df.head()

prod_features.dtypes

user_features.dtypes
